+++ integreat_cms/cms/models/languages/language.py
@@ -97,7 +97,7 @@ class Language(AbstractBaseModel):
             _("This is used in exported PDFs."),
         ),
     )
-    message_live_content_not_available = models.CharField(
+    message_content_not_available = models.CharField(
         max_length=250,
         blank=False,
         default="This page does not exist in the selected language. It is however available in these languages:",
@@ -105,7 +105,7 @@ class Language(AbstractBaseModel):
             '"This page does not exist in the selected language. It is however available in these languages:" in this language'
         ),
         help_text=_(
-            "This is shown to the user when a mirrored page is not available in their language."
+            "This is shown to the user when a page is not available in their language."
         ),
     )
     message_partial_live_content_not_available = models.CharField(