+++ integreat_cms/cms/models/pois/poi.py
@@ -1,4 +1,5 @@
 from django.db import models
+from django.core.validators import MaxValueValidator, MinValueValidator
 from django.utils.translation import ugettext_lazy as _
 
 from ..abstract_content_model import AbstractContentModel
@@ -19,15 +20,23 @@ class POI(AbstractContentModel):
     city = models.CharField(max_length=250, verbose_name=_("city"))
     country = models.CharField(max_length=250, verbose_name=_("country"))
     latitude = models.FloatField(
-        verbose_name=_("latitude"), help_text=_("The latitude coordinate")
+        null=True,
+        blank=True,
+        verbose_name=_("latitude"),
+        help_text=_("The latitude coordinate"),
+        validators=[MinValueValidator(-90.0), MaxValueValidator(90.0)],
     )
     longitude = models.FloatField(
-        verbose_name=_("longitude"), help_text=_("The longitude coordinate")
+        null=True,
+        blank=True,
+        verbose_name=_("longitude"),
+        help_text=_("The longitude coordinate"),
+        validators=[MinValueValidator(-180.0), MaxValueValidator(180.0)],
     )
-    location_not_on_map = models.BooleanField(
-        default=False,
-        verbose_name=_("Do not show this location on map"),
-        help_text=_("Tick if you do not show this location on map"),
+    location_on_map = models.BooleanField(
+        default=True,
+        verbose_name=_("Show this location on map"),
+        help_text=_("Tick if you want to show this location on map"),
     )
     icon = models.ForeignKey(
         MediaFile,
@@ -40,6 +49,14 @@ class POI(AbstractContentModel):
         default=False,
         verbose_name=_("archived"),
         help_text=_("Whether or not the location is read-only and hidden in the API."),
+    )
+    website = models.URLField(max_length=250, blank=True, verbose_name=_("website"))
+    email = models.EmailField(
+        blank=True,
+        verbose_name=_("email address"),
+    )
+    phone_number = models.CharField(
+        max_length=250, blank=True, verbose_name=_("phone number")
     )
 
     @staticmethod