+++ integreat_cms/static/src/editor.ts
@@ -6,3 +6,4 @@
 import './js/forms/tinymce-init.ts';
 import './js/pages/sbs-copy-content.ts';
 import "./js/content-edit-lock.ts";
+import "./js/pages/page-preview-form.ts";