+++ integreat_cms/static/src/js/pages/fetch-subpages.ts
@@ -8,29 +8,31 @@ import { setBulkActionEventListeners } from "../bulk-actions";
 import { setToggleSubpagesEventListeners } from "./toggle-subpages";
 import { addDragAndDropListeners } from "../tree-drag-and-drop";
 import { addConfirmationDialogListeners } from "../confirmation-popups";
+import { addPreviewWindowListeners, openPreviewWindowInPageTree } from "./page-preview";
 
 window.addEventListener("load", () => {
-  // load subpages initially
+  // Load subpages initially
   if (document.querySelector("[data-delay-event-handlers]")) {
     let rootPages: HTMLElement[] = Array.from(document.querySelectorAll(".toggle-subpages"));
     console.debug("Loading all subpages");
     Promise.all(rootPages.map(fetchSubpages)).then(() => {
       console.debug("Finished loading all subpages");
-      // render icons for recently added DOM elements
+      // Render icons for recently added DOM elements
       feather.replace({ class: 'inline-block' });
       // Set event handlers
       setToggleSubpagesEventListeners();
       setBulkActionEventListeners();
       addDragAndDropListeners();
       addConfirmationDialogListeners();
+      addPreviewWindowListeners(openPreviewWindowInPageTree);
     })
   }
 });
 
 /**
- * Ajax call to fetch children of selected page 
+ * Ajax call to fetch children of selected page
  * and insert them into DOM at right position
- * 
+ *
  * @param collapseSpan The page to expand its children
  */
 export async function fetchSubpages(collapseSpan: HTMLElement): Promise<number[]> {
@@ -55,7 +57,7 @@ export async function fetchSubpages(collapseSpan: HTMLElement): Promise<number[]
   let descendantIds: number[] = [];
   renderedDescendants.forEach(function(rowToInsert: HTMLTableRowElement) {
     if (rowToInsert.classList.contains("page-row")) {
-      // record children ids to update all ancestors descendants data attribute
+      // Record children ids to update all ancestors descendants data attribute
       let descendantId = parseInt(rowToInsert.dataset.dropId);
       descendantIds.push(descendantId);
       // Check if the descendant is a direct child
@@ -63,9 +65,9 @@ export async function fetchSubpages(collapseSpan: HTMLElement): Promise<number[]
         directChildrenIds.push(descendantId);
       }
     }
-    // insert child into DOM tree
+    // Insert child into DOM tree
     currentRow.after(rowToInsert);
-    // update point to insert next child
+    // Update point to insert next child
     currentRow = rowToInsert;
   });
   // Update descendants to enable drag & drop functionality