+++ tests/cms/views/view_config.py
@@ -303,7 +303,7 @@ VIEWS = [
                     "text_direction": "LEFT_TO_RIGHT",
                     "table_of_contents": "Inhaltsverzeichnis",
                     "primary_country_code": "de",
-                    "message_live_content_not_available": "Foo1",
+                    "message_content_not_available": "Foo1",
                     "message_partial_live_content_not_available": "Foo2",
                 },
             ),
@@ -408,7 +408,6 @@ VIEWS = [
     ),
     (
         [
-            ("view_page", STAFF_ROLES + [MANAGEMENT, EDITOR]),
             ("edit_page", STAFF_ROLES + [MANAGEMENT, EDITOR]),
             ("edit_page", STAFF_ROLES + [MANAGEMENT, EDITOR]),
             (
@@ -441,7 +440,6 @@ VIEWS = [
     ),
     (
         [
-            ("view_page", STAFF_ROLES),
             ("edit_page", STAFF_ROLES),
             ("sbs_edit_page", STAFF_ROLES),
             ("page_revisions", STAFF_ROLES),